Sharing Doernerian emotional model between engines
