+++ micropsi_core/nodenet/dict_engine/dict_nodenet.py
@@ -8,7 +8,8 @@ import micropsi_core
 from micropsi_core.nodenet import monitor
 from micropsi_core.nodenet.node import Nodetype
 from micropsi_core.nodenet.nodenet import Nodenet, NODENET_VERSION, NodenetLockException
-from .dict_stepoperators import DictPropagate, DictPORRETDecay, DictCalculate, DictDoernerianEmotionalModulators
+from micropsi_core.nodenet.stepoperators import DoernerianEmotionalModulators
+from .dict_stepoperators import DictPropagate, DictPORRETDecay, DictCalculate
 from .dict_node import DictNode
 from .dict_nodespace import DictNodespace
 import copy
@@ -187,7 +188,7 @@ class DictNodenet(Nodenet):
 
         super(DictNodenet, self).__init__(name, worldadapter, world, owner, uid)
 
-        self.stepoperators = [DictPropagate(), DictCalculate(), DictPORRETDecay(), DictDoernerianEmotionalModulators()]
+        self.stepoperators = [DictPropagate(), DictCalculate(), DictPORRETDecay(), DoernerianEmotionalModulators()]
         self.stepoperators.sort(key=lambda op: op.priority)
 
         self.__version = NODENET_VERSION  # used to check compatibility of the node net data
